Map ICMP to ICMP6 in REJECT's reject-with
Ferm currently allows the use of "icmp-type" in ip6 domains and
automatically maps it to icmp6-type.

That's not the case for REJECT's target --reject-with, which would be
especially useful for constructs like:

    domain (ip ip6) chain INPUT {
        proto tcp dport 22 REJECT reject-with icmp-admin-prohibited;
    }

Implement the above and try to also be (too?) smart and map all the
possible iptables arguments to their respective ip6tables, which is not
always a 1:1 mapping.