Make @resolve do the right thing on domain ip6
Currently @resolve, when not passed an RRTYPE, defaults to A,
irrespective of the domain. This causes code such as:
  domain ip6 table filter chain INPUT saddr @resolve(example.org) ACCEPT;
...to output an ip6tables rule with an IPv4 address.

Moreover, this statement:
  domain (ip ip6) table filter chain INPUT saddr @resolve(example.org) ACCEPT;
...currently returns on iptables rule with the IPv4 address, and no IPv6
rules at all.

Instead, make @resolve's default RRTYPE be conditional to the domain, A
for ip (as before) and AAAA on domain ip6. This makes both of the above
examples work and do what would be intuitively expected from such
statements.