Fix and enable dual-stack checks in the test suite
While executing the test suite in fast mode, both import-ferm as well as
the output diff in dual-stack configurations is currently broken.
Because of that, the following tests were currently explicitly disabled:
* ipv6/mixed.ferm
* misc/address-magic.ferm
* misc/ipfilter.ferm
* misc/subchain-domains.ferm

To fix this, first add the tool name (e.g. iptables-save) in the
"generated by" marker at the top of the fast mode output. The marker now
looks like this:
  Generated by ferm 2.4.2~git (iptables-save) on Fri Oct 26 19:54:12 2018

import-ferm already had regexps to recognize the tool name and set the
domain, and thus is now able to understand ferm's own output. This also
allows us to drop the explicit FERM_DOMAIN=ip6 setting in the Makefile.

Finally, modify test/sort.pl to handle this in two ways:
* not skip this marker in the output (but skip the date);
* sort/emit each domain's rules separately.