Use perl from $PATH, not /usr/bin
This is currently broken for example on Travis, which uses Perl from
perlbrew in order to provide testing in multiple versions etc. While our
Travis config dictates testing with a couple of different Perl versions,
in reality only the system's Perl in /usr/bin is used.

Switch to just the relative "perl" and let $PATH do its magic.