Fix import-ferm non-optimized code with Perl 5.18+
import-ferm tries to optimize the code by trying to group statements
into blocks and rules into arrays. To do that, it eliminates members
from hashes, and then compares the Data::Dumper of the remainders.

Unfortunately, starting with Perl 5.18, the order of hashes is
randomized, and so these optimizations are now happening at random. Set
$Data::Dumper::Sortkeys to 1 to restore sanity to this test, and add
Perl 5.18 (and while at it, 5.26) to the Travis-tested versions to make
this more evident.