test: use --test-mock-previous for the "preserve" tests
Using network namespaces for the unit tests is a clever idea, but
unfortunately it is very fragile.  It breaks if the kernel doesn't
have iptables support or if the iptables tools are not installed, and
requires root privileges.

Instead of going through all those hoops, let's just mock the previous
iptables ruleset.